requirements.txt
django==3.1
chatbotai==0.3.0.0
django-chatbot==0.0.2.2
messengerbot==0.1.4
requests==2.22.0
wikipedia==1.4.0

